Fahrzeug löschen Bug fixed

+++ app/src/main/java/com/example/tankauswertung/Garage.java
@@ -166,9 +166,11 @@ public class Garage {
                 ArrayList<Tankvorgang> tankvorgaenge = f.getTankvorgaenge();
                 for (int i = 0; i < tankvorgaenge.size(); i++) {
                     Tankvorgang t = tankvorgaenge.get(i);
-                    File file = new File(t.getImg());
-                    if (!file.delete()) {
-                        System.out.println("Bild: \"" + t.getImg() + "\" konnte nicht geloescht werden" );
+                    if (t.getImg() != null) {
+                        File file = new File(t.getImg());
+                        if (!file.delete()) {
+                            System.out.println("Bild: \"" + t.getImg() + "\" konnte nicht geloescht werden" );
+                        }
                     }
                 }
                 fahrzeuge.remove(key);
